requirements-clap.txt
git+https://github.com/LAION-AI/CLAP.git@clap
torchaudio
h5py
torchlibrosa
PySoundFile